# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.LoginResponse:
    properties:
      access_token:
        type: string
      company_id:
        type: string
      email:
        type: string
      expires_in:
        description: seconds
        type: integer
      roles:
        items:
          type: string
        type: array
      token_type:
        description: '"Bearer"'
        type: string
      user_id:
        type: string
    type: object
  auth.MeResponse:
    properties:
      company_id:
        type: string
      email:
        type: string
      roles:
        items:
          type: string
        type: array
      user_id:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      company_id:
        type: string
      email:
        type: string
      password:
        type: string
      role:
        description: '"admin" or "user"'
        type: string
    type: object
  auth.RegisterResponse:
    properties:
      company_id:
        type: string
      email:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
...
  filemeta.CreateFileMetaRequest:
    properties:
      company_id:
        description: can be nil if not tied to a company
        type: string
      file_key:
        description: required (e.g. S3 key)
        type: string
      file_name:
        type: string
      file_size:
        description: required
        type: integer
      file_txn_meta:
        type: string
      file_txn_type:
        description: required (e.g. 1=upload,2=delete, etc.)
        type: integer
    type: object
  filemeta.CreateFileMetaResponse:
    properties:
      id:
        type: string
    type: object
  filemeta.FileMetaResponse:
    properties:
      company_id:
        type: string
      created_at:
        type: string
      file_key:
        type: string
      file_name:
        type: string
      file_size:
        type: integer
      file_txn_meta:
        type: string
      file_txn_type:
        type: integer
      id:
        type: string
    type: object
...
host: localhost:9393
info:
  contact: {}
  description: API for Jupyter Platform (DMS, contact-us, etc).
  title: Jupyter API
  version: 1.0.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Validates credentials and returns a JWT access token
      parameters:
      - description: Login payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: invalid request
          schema:
            type: string
        "401":
          description: invalid credentials
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Login with email and password
      tags:
      - auth
  /auth/me:
    get:
      description: Returns user_id, company_id, email and roles from the access token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.MeResponse'
        "401":
          description: unauthorized
          schema:
            type: string
      summary: Get current user info from JWT
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a user with Argon2 password hash
      parameters:
      - description: Register payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: invalid request
          schema:
            type: string
        "409":
          description: email already exists
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Register a new user
      tags:
      - auth
...
  /files/meta:
    post:
      consumes:
      - application/json
      description: Stores metadata for a file transaction (e.g. upload to S3)
      parameters:
      - description: File meta payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/filemeta.CreateFileMetaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/filemeta.CreateFileMetaResponse'
        "400":
          description: invalid request
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Create file metadata
      tags:
      - files
  /files/meta/{id}:
    get:
      description: Fetch a single files_meta row by its ID
      parameters:
      - description: File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/filemeta.FileMetaResponse'
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: internal error
          schema:
            type: string
      summary: Get file meta by ID
      tags:
      - files
...
swagger: "2.0"